Many fix,,such refactor
