+++ blaseball_stlats_tlracker/__init__.py
@@ -5,3 +5,6 @@ blaseball_stlats_tlracker
 # Reference:  https://betterscientificsoftware.github.io/python-for-hpc/tutorials/python-pypi-packaging/
 
 __version__ = "0.1"
+
+# from <.py file name> import <class name>
+from BST import BST
